responses.txt
The difference between the Web and the internet is that the web is a big collection of information that is stored and ready for us to use it when we want. The internet is what helps us connect to the web and get access to the information that we are looking for. The internet works by taking data and condensing it down into packets. The packets are sent through the vast networks that connect one computer to the rest of the world.
A directory is basically a file that holds all relevant references that may be needed. VCS is a great way to track and save progress as you are going that helps you keep older versions of a project so that way you'll have that to go back to and reference if need be.